Add plugin version to GRPC interface
Unlike the POC, I went ahead and added it into the main interface. So
far in my testing, this has been safe to do.

I also took a slightly different approach for how to include the
`v1.12.0+builtin` version in the builtin plugins: I added a new wrapper
struct that sets the version to the "builtin" version if it isn't
already set, and I wrap all of the builtin factories with that when we
create the map of factories. This way we don't have to update every
builtin plugin to add the version logic. If we do want to version those
plugins separately, though, we will be able to do so.

There are still some TODOs before merging this, I feel, but this is
getting large so I wanted to check in:

* [ ] More manual testing with existing plugins and updated (with
  version) ones
* [ ] Update identity/cubbyhole/sys/etc. special backends to support
  version method.