Allow patching a role to create a new one
When PKI's role patch operation was first created, we didn't allow
creating a role if it didn't exist. Now that a patch-capable CLI is
supported, users might wish to use that instead for high-impact
endpoints like role modification.

Support creating roles via patch operation, as per the RFC 5789, it
leaves it to us to decide:

    The PATCH method requests that a set of changes described in the
    request entity be applied to the resource identified by the Request-
    URI.  The set of changes is represented in a format called a "patch
    document" identified by a media type.  If the Request-URI does not
    point to an existing resource, the server MAY create a new resource,
    depending on the patch document type (whether it can logically modify
    a null resource) and permissions, etc.

Signed-off-by: Alexander Scheel <alex.scheel@hashicorp.com>